Prototype fix for PR44147
Value pointers in LazyValueInfos collection of over-defined Values can be
freed then reallocated, but remain in the set. This leads to spurious
reports of over-defined Values, that happen non-deterministically according
due to varying pointer locations.

I haven't looked closely at whether more adjacent code needs this fix, or
whether there are performance implications, but it reduces the symptoms of
PR44147.