Commit some half-baked register-tracking stuff
This is half done, note the liberal use of abort and __asm__("int $3").
This now successfully tracks the elimination of identity copies (in
VirtRegRewriter) back to the defining inst. This leads to a situation
where we have to track implicit register defs from things like call
instructions: they don't have an operand for the return value, but we
have to describe the value that they define in some way.