Kill off the vulnerable copies idea
I've had a better one: lets just back-propagate all copies before regalloc
happens. That way, we can find a vreg at the start of a block that's acting
as a PHI, and look it up at that position. The alternative is a COPY
becoming a stack spill, after which it's difficult to map it back to vregs.