Strengthen testing for backedge vloc joining
If we produce a vphi, then any path without a DBG_VALUE on it will join
successfully, and the transfer-tracker will emit a location there just
fine. However, this leaves us vulnerable to backedge paths where the
variable location isn't modified, but it doesn't actually reach the point
of the mphi.