Override lots of reg locations on spill/restore
Specifically, we clobber all aliases when restoring, and only track the
subreg that we assign to. Also def regs after spilling: we end up with
multiple locations for them otherwise, which is painful. And there are
cases where there are spills that must be tracked, then reads from the
spilt reg that must be prop'd.