Switch over to new LDV implementation.
This was developed in a different branch to ensure that there are zero
changes between this and a normal LiveDebugValues... this file is me
copying the version from a1e26139308cc5a5fe7c80c8b513845b11bb6dc9 and
shoehorning in the instr ref stuff. Will have to massage in other
patches concurrently.