Correct a liveness query for regcoalescing
We copy into the PHI at the start of the block; but not necessarily _by_
the start of the block. Try to find that copy and pick its slot index,
otherwise the liveness query will be out of range (too early).

This could be done better, but phi eliminations main function is hard,
ignore it for now and go for an approximate solution