Spray subregisters everywhere
This is, well, pretty hard to implement due to Design Decisions. It
would be oh so much better if we didn't have to do so much work mangling
copies in register coalescing, however, we're still not at a point of
avoiding that. The world would be a better place if we were.