Ignore this
Oh look: Tail Duplication actaully peels apart the SSA-form of debug locs
that I was trying to preserve; and quite often during block placement. This
hugely sucks; but is what happens.

To get around this, record the transform that was occurring, and then piece
together the SSA form later in LiveDebugValues. Probably highly expensive; but
then again, that's what's accurate.