Delete a check for an impossible scenario
This has no effect on a clang-3.4 build. Here, we're converting a
ValueIDNum into a location, and if the incoming value is an mphi from this
block, we use the mphi definition location.

However, if we couldn't find that in OLoc's incoming locations, then this
will never be a legitimate merge anyway. Delete this pointless test.