Switch to a better implicit def solution
In this scenario:

  entry:
    call @foo
    call @bar
    $rdx = IMPLICIT_DEF
    br someblock

There's no explicit write to rdx anywhere, so we're not tracking it. However,
the register is explicitly set to be live there, for reasons known only to
the register allocator. In this scenario, if there's also a DBG_VALUE of
$rdx somewhere, it needs to read something meaningful, or we'll fail to
join variable locations later. Any empty location (LocIdx == 0) seen on
joins is considered a failure.

To work around this, if there isn't a value for the IMPLICIT_DEF'd register
yet, then explicitly def it.