Add a switch for choosing what variable location tracker to use
This CodeGen / Xclang switch currently chooses which implementation of
LiveDebugValues to use. Over time however, we'll be adding fragments of
support in aid of this RFC:

  http://lists.llvm.org/pipermail/llvm-dev/2020-March/139636.html

Which will slowly move variable locations to be defined by an instruction
that calculates a value, and a DBG_INSTR_REF instruction referring to that
value. Thus, this option will grow into "use the new kind of variable
locations" rather than "use the new LiveDebugValues" implementation.