[WIP][DebugInstrRef][2/9] Parse debug instruction-references from/to MIR
This patch sets a MIR format for debug instruction references: it's an integer
trailing an instruction, marked out by "debug-instr-number", much like how
"debug-location" identifies the DebugLoc metadata of an instruction. I'm not
wed to this name, or particularly to this syntax, but something is needed.

I've added a round-trip test and two verifier checks: that we don't label
meta-instructions as generating values, and that there are no duplicates. Note
that the MachineFunction count of which instruction number should be used next
is produced by just picking the highest instruction number and adding one.