[WIP][DebugInstrRef][5/9] Substitute debug value numbers to handle optimisations
This patch handles two optimisations, TwoAddressInstruction and X86's FixupLEAs
pass, both of which optimise by re-creating instructions.  For LEAs, various
bits of arithmetic are better represented as LEAs on X86, while
TwoAddressInstruction sometimes converts instrs into three address instructions
if it's profitable. Both of these require substitutions to be created -- test
that an entry is made in the MIR output. (Handling of this in LiveDebugValues
comes later).