[WIP][DebugInstrRef][7/9] NFC: Separate collection of machine/variable values
This patch has InstrRefBasedLDV step through the entire function again, looking
for DBG_VALUE instructions, after the machine value number problem has been
solved. This is what an early implementation of mine used to do; resolving
variable values into machine value numbers in vlocJoin was a late optimisation
I introduced.

The reason for changing back now, is that when we find a DBG_INSTR_REF, it's
very convenient to already have stepped through the function once already, so
that we can immediately determine what machine value number it refers to. This
makes the next patch much easier to write.

(This patch makes things slower; I'm seeing that it'll be easier to review in
the simple-slow case, then add optimisation later).